pyproject.toml
[project]
name = "quant-trading"
version = "0.1.0"
description = "Quant trading project bootstrap"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "sqlalchemy",
    "pymysql",
    "lightgbm",
    "pyqlib",
    "quantstats",
    "tushare",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
qt = "quant_trading.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
